+++ lucene/CHANGES.txt
@@ -18,6 +18,9 @@ API Changes
 * GITHUB#12592: Add RandomAccessInput#length method to the RandomAccessInput interface. In addition deprecate
   ByteBuffersDataInput#size in favour of this new method. (Ignacio Vera)
 
+* GITHUB#12646: Move FST#addNode to FSTCompiler to avoid a circular dependency between FST and FSTCompiler
+  (Anh Dung Bui via Alan Woodward)
+
 New Features
 ---------------------
 (No changes)
